Update aggregate GitHub action to push to completions branch (#21)
The main branch is protected and needs a pull request for a change to be
made. This cannot be bypassed by GitHub actions. An unhelpful solution
from GitHub is to create a GitHub app which is overkill. For now,
completions has no protections which is potentially a security hole.

### Changed
* Aggregate GitHub action to push to completions branch

Tracked against: None